requirements.txt
streamlit
pandas
openpyxl
pydantic
python-dotenv
cryptography
rapidfuzz
